feat(selection): open "my selection" as a modal over the menu; reuse QR intent (#15)
Task 1: extract the selection UI into a shared SelectionContent component used
by both the /selection page (chrome + back link) and a new bottom-sheet Dialog
on the menu. The menu pill is now a button that opens the modal (no navigation,
menu stays mounted behind); staff add-order still deep-links to /selection.

Task 2: cache the created order intent (url + expiresAt + line fingerprint);
reopening the QR with an unchanged cart and >2min left reuses it instead of
POSTing again, so a diner toying with the button no longer burns the 10/min
per-IP limit. Added a 429-specific selection.qrRateLimit string in all 10 locales.

Tests: extended SelectionPageClient unit tests (intent reuse, 429 error); added
a MenuPageClient modal-open test; updated e2e specs for the pill-as-button and a
no-navigation assertion. web+backend+chat unit suites green, lint clean,
i18n-completeness green, full Playwright suite green (119).
